Split C++ core, remove Rust implementation
