+++ .gitignore
@@ -1,4 +1,3 @@
-/target
 /rel
 *.png
 *.gif
