+++ CMakeLists.txt
@@ -3,7 +3,12 @@ project(falsecolor)
 
 find_package(Python3 REQUIRED COMPONENTS Interpreter Development NumPy)
 
-add_library(falsecolor SHARED ${CMAKE_CURRENT_LIST_DIR}/falsecolor.cpp)
+add_library(falsecolor SHARED
+    ${CMAKE_CURRENT_LIST_DIR}/src/falsecolor.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/src/fit.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/src/image.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/src/smudge.cpp
+)
 
 if(NOT Python3_NumPy_FOUND)
     message(FATAL_ERROR "NumPy headers not found")
